+++ lib/zoi.ex
@@ -64,16 +64,19 @@ defmodule Zoi do
   @doc group: "Parsing"
   @spec parse(schema :: Zoi.Type.t(), input :: input(), opts :: options) :: result()
   def parse(schema, input, opts \\ []) do
+    ctx = Keyword.get(opts, :ctx, Zoi.Context.new(schema, input))
+    opts = Keyword.put_new(opts, :ctx, ctx)
+
     with {:ok, result} <- Zoi.Type.parse(schema, input, opts),
-         {:ok, result} <- Meta.run_transforms(schema, result),
-         {:ok, _refined_result} <- Meta.run_refinements(schema, result) do
+         ctx = Zoi.Context.add_parsed(ctx, result),
+         {:ok, result} <- Meta.run_transforms(ctx),
+         ctx = Zoi.Context.add_parsed(ctx, result),
+         {:ok, _refined_result} <- Meta.run_refinements(ctx) do
       {:ok, result}
     else
-      {:error, reason} when is_binary(reason) ->
-        {:error, Zoi.Errors.add_error(reason)}
-
       {:error, error} ->
-        {:error, error}
+        ctx = Zoi.Context.add_error(ctx, error)
+        {:error, ctx.errors}
     end
   end
 
@@ -898,20 +901,57 @@ defmodule Zoi do
     |> transform({Zoi.Transforms, :transform, [[:to_upcase]]})
   end
 
-  @doc """
+  @doc ~S"""
   Adds a custom validation function to the schema.
 
   This function will be called with the input data and options, and should return `:ok` for valid data or `{:error, reason}` for invalid data.
 
-  ## Example
-
-      iex> schema = Zoi.string() |> Zoi.refine(fn _schema, value -> 
+      iex> schema = Zoi.string() |> Zoi.refine(fn value -> 
       ...>   if String.length(value) > 5, do: :ok, else: {:error, "must be longer than 5 characters"}
       ...> end)
       iex> Zoi.parse(schema, "hello")
       {:error, [%Zoi.Error{message: "must be longer than 5 characters"}]}
       iex> Zoi.parse(schema, "hello world")
       {:ok, "hello world"}
+
+  ## Returning multiple errors
+
+  You can use the context when defining the `Zoi.refine/2` function to return multiple errors.
+
+      iex> schema = Zoi.string() |> Zoi.refine(fn value, ctx ->
+      ...>   if String.length(value) < 5 do
+      ...>     Zoi.Context.add_error(ctx, "must be longer than 5 characters")
+      ...>     |> Zoi.Context.add_error("must be shorter than 10 characters")
+      ...>   end
+      ...> end)
+      iex> Zoi.parse(schema, "hi")
+      {:error, [
+        %Zoi.Error{message: "must be longer than 5 characters"},
+        %Zoi.Error{message: "must be shorter than 10 characters"}
+      ]}
+
+  ## mfa 
+
+  You can also pass a `mfa` (module, function, args) to the `Zoi.refine/2` function. This is recommended if
+  you are declaring schemas during compile time:
+
+      defmodule MySchema do
+        use Zoi
+
+        @schema Zoi.string() |> Zoi.refine({__MODULE__, :validate, []})
+
+        def validate(value, opts \\ []) do
+          if String.length(value) > 5 do
+            :ok
+          else 
+            {:error, "must be longer than 5 characters"}
+          end
+        end
+      end
+
+  Since during the module compilation, anonymous functions are not available, you can use the `mfa` option to pass a module, function and arguments.
+  The `opts` argument is mandatory, this is where the `ctx` is passed to the function and you can leverage the `Zoi.Context` to add extra errors.
+  In general, most cases the `:ok` or `{:error, reason}` returns will be enough. Use the context only if you need extra errors or modify the context in some way.
   """
   @doc group: "Extensions"
   @spec refine(schema :: Zoi.Type.t(), fun :: refinement()) :: Zoi.Type.t()
@@ -946,7 +986,7 @@ defmodule Zoi do
 
   ## Example
 
-      iex> schema = Zoi.string() |> Zoi.transform(fn _schema, value ->
+      iex> schema = Zoi.string() |> Zoi.transform(fn value ->
       ...>   {:ok, String.trim(value)}
       ...> end)
       iex> Zoi.parse(schema, "  hello world  ")
@@ -972,7 +1012,6 @@ defmodule Zoi do
     %{schema | schemas: schemas}
   end
 
-  @spec transform(schema :: Zoi.Type.t(), fun :: function()) :: Zoi.Type.t()
   def transform(schema, fun) do
     update_in(schema.meta.transforms, fn transforms ->
       transforms ++ [fun]